Adding all the hospitals in the seed

+++ bulk_fhir_resource_curator/src/main/resources/db/migration/V4__CreateObservationTableTable.sql
@@ -3,7 +3,7 @@ CREATE TABLE observation(
     patient_id VARCHAR(255) NOT NULL,
     issued DATE NOT NULL,
     code_text VARCHAR(255) NOT NULL,
-    quantity_value BIGINT NOT NULL,
+    quantity_value DOUBLE NOT NULL,
     quantity_unit VARCHAR(32) NOT NULL,
     source_system_id BIGINT NOT NULL,
     source_system_name VARCHAR(255) NOT NULL,
